+++ CHANGELOG.md
@@ -20,6 +20,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - `YoloAIO.py`: Reworked CLI UI to behave more intuitively.
 -->
 
+## [0.4.dev1] - 2021-11-09
+
+### Added
+
+- `DRSlib.decorators`: Added `call_progress`
+
+### Modified
+
+- `DRSlib.debug`: Module docstring cleanup.
+
 ## [0.3] - 2021-11-08
 
 ### Added